+++ lib/src/commands/check_command.dart
@@ -1,4 +1,4 @@
-import 'package:cura/src/commands/base_command.dart';
+import 'package:cura/src/commands/base/base_command.dart';
 import 'package:cura/src/domain/models/package_health.dart';
 import 'package:cura/src/presentation/formatters/table_formatter.dart';