+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Config Hierarchy - Global vs Project Config
+
 ## [0.2.0] - 2026-02-13
 
 ### Added
 
-- Config Hierarchy - Global vs Project Config
 - add README documentation with base file in ./docs
 - configuration system support
 - complete theme system with dark/light mode
@@ -27,5 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Operational local cache.
 - ScoreCalculator unit tests (>80% coverage)
 
-[unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.0...HEAD
-[0.1.0]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.0.1
+[unreleased]: https://github.com/meragix/cura/compare/cura-0.2.0...HEAD
+[0.2.0]: https://github.com/meragix/cura/releases/tag/cura-0.2.0
+[0.1.0]: https://github.com/meragix/cura/releases/tag/cura-0.1.0